// src/renderer/types.ts
export interface Property {
  courage?: PropertyValue;
  wisdom?: PropertyValue;
  intuition?: PropertyValue;
  charisma?: PropertyValue;
  fingerAbility?: PropertyValue;
  dexterity?: PropertyValue;
  constitution?: PropertyValue;
  strength?: PropertyValue;
  socialStatus?: PropertyValue;
  lifeforce?: PropertyValue;
  endurance?: PropertyValue;
  astralEnergy?: PropertyValue;
  karmaEnergy?: PropertyValue;
  magicResistance?: PropertyValue;
  initiativBaseValue?: PropertyValue;
  attackBaseValue?: PropertyValue;
  paradeBaseValue?: PropertyValue;
  remoteCombatBaseValue?: PropertyValue;
  risk?: PropertyValue;
  loyalty?: PropertyValue;
  speed?: PropertyValue;
  magicResistance2?: PropertyValue;
  armor?: PropertyValue;
  [key: string]: PropertyValue | undefined;
}

interface PropertyValue {
  mod: number;
  value: number;
  basicValue: number;
  greatMeditation?: number;
  magicResistenceMod?: number;
  startValue?: number;
  calcValue?: number;
}